% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lifeexpectancy.R
\name{nationchoice}
\alias{nationchoice}
\title{Selection the country to plot data}
\usage{
nationchoice(nation)
}
\arguments{
\item{nation}{List of Asian Pacific countries}
}
\value{
result plot the data of the selected country
}
\description{
Creates a plot of life expectancy of country was choosen in the list nation
}
\details{
Give the selection on 45 countries of Asian Pacific within the nation list to plot data of life expectancy
}
\note{
Please, check the list of countries in the nation list and pay attention on capital letter
}
\examples{
nationchoice("Vietnam")
}
